+++ devito/__init__.py
@@ -12,7 +12,7 @@ from devito.data import *  # noqa
 from devito.dimension import *  # noqa
 from devito.equation import *  # noqa
 from devito.finite_differences import *  # noqa
-from devito.function import Buffer, NODE, CELL  # noqa
+from devito.functions.utils import Buffer, NODE, CELL  # noqa
 from devito.logger import error, warning, info, set_log_level, silencio  # noqa
 from devito.parameters import *  # noqa
 from devito.tools import *  # noqa
