+++ devito/dse/backends/basic.py
@@ -7,7 +7,7 @@ from devito.dse.backends import AbstractRewriter, dse_pass
 from devito.dse.manipulation import common_subexprs_elimination
 from devito.symbolics import (Eq, bhaskara_cos, bhaskara_sin, retrieve_indexed,
                               q_affine, q_scalar)
-from devito.types import Scalar
+from devito.functions.types import Scalar
 
 
 class BasicRewriter(AbstractRewriter):
