+++ devito/functions/__init__.py
@@ -1,0 +1,5 @@
+from .types import *  # noqa
+from .dense import * # noqa
+from .sparse import *  # noqa
+from .constants import *  # noqa
+from .utils import *  # noqa
