+++ devito/functions/types.py
