+++ devito/ir/iet/nodes.py
@@ -20,7 +20,7 @@ from devito.dimension import Dimension
 from devito.symbolics import FunctionFromPointer, as_symbol
 from devito.tools import (Signer, as_tuple, filter_ordered, filter_sorted, flatten,
                           validate_type)
-from devito.types import AbstractFunction, Symbol, Indexed
+from devito.functions.types import AbstractFunction, Symbol, Indexed
 
 __all__ = ['Node', 'Block', 'Denormals', 'Expression', 'Element', 'Callable',
            'Call', 'Conditional', 'Iteration', 'List', 'LocalExpression',
